+++ rust/consensus/Cargo.toml
@@ -9,4 +9,5 @@ tokio = { version = "1.5.0", features = ["sync"] }
 log = "0.4.14"
 
 crypto = { path = "../crypto" }
+config = { path = "../config" }
 primary = { path = "../primary" }
