+++ rust/primary/Cargo.toml
@@ -6,7 +6,6 @@ publish = false
 edition = "2018"
 
 [dependencies]
-displaydoc = "0.1.7"
 failure = "0.1.8"
 rstest = "0.6.4"
 rand = "0.7.3"
@@ -24,6 +23,10 @@ bytes = "1.0.1"
 async-std = { version = "1.5", features = ["attributes"] }
 env_logger = "0.7.1"
 log = "0.4.11"
+async-recursion = "0.3.2"
+async-trait = "0.1.50"
 
-store = { path = "../store" }
 crypto = { path = "../crypto" }
+store = { path = "../store" }
+config = { path = "../config" }
+network = { path = "../network" }
