+++ rust/primary/src/aggregators.rs
@@ -1,0 +1,84 @@
+use crate::error::{DagError, DagResult};
+use crate::messages::{Certificate, Header, Vote};
+use config::{Committee, Stake};
+use crypto::Hash as _;
+use crypto::{Digest, PublicKey, Signature};
+use std::collections::HashSet;
+
+/// Aggregates votes for a particular header into a certificate.
+pub struct VotesAggregator {
+    weight: Stake,
+    votes: Vec<(PublicKey, Signature)>,
+    used: HashSet<PublicKey>,
+}
+
+impl VotesAggregator {
+    pub fn new() -> Self {
+        Self {
+            weight: 0,
+            votes: Vec::new(),
+            used: HashSet::new(),
+        }
+    }
+
+    pub fn append(
+        &mut self,
+        vote: Vote,
+        committee: &Committee,
+        header: &Header,
+    ) -> DagResult<Option<Certificate>> {
+        let author = vote.author;
+
+        // Ensure it is the first time this authority votes.
+        ensure!(self.used.insert(author), DagError::AuthorityReuse(author));
+
+        self.votes.push((author, vote.signature));
+        self.weight += committee.stake(&author);
+        if self.weight >= committee.quorum_threshold() {
+            self.weight = 0; // Ensures quorum is only reached once.
+            return Ok(Some(Certificate {
+                header: header.clone(),
+                round: vote.round,
+                origin: vote.origin,
+                votes: self.votes.clone(),
+            }));
+        }
+        Ok(None)
+    }
+}
+
+/// Aggregate certificates and check if we reach a quorum.
+pub struct CertificatesAggregator {
+    weight: Stake,
+    certificates: Vec<Digest>,
+    used: HashSet<PublicKey>,
+}
+
+impl CertificatesAggregator {
+    pub fn new() -> Self {
+        Self {
+            weight: 0,
+            certificates: Vec::new(),
+            used: HashSet::new(),
+        }
+    }
+
+    pub fn append(
+        &mut self,
+        certificate: Certificate,
+        committee: &Committee,
+    ) -> DagResult<Option<Vec<Digest>>> {
+        let origin = certificate.origin;
+
+        // Ensure it is the first time this authority votes.
+        ensure!(self.used.insert(origin), DagError::AuthorityReuse(origin));
+
+        self.certificates.push(certificate.digest());
+        self.weight += committee.stake(&origin);
+        if self.weight >= committee.quorum_threshold() {
+            self.weight = 0; // Ensures quorum is only reached once.
+            return Ok(Some(self.certificates.drain(..).collect()));
+        }
+        Ok(None)
+    }
+}
