+++ rust/primary/src/core.rs
@@ -1,0 +1,366 @@
+use crate::aggregators::{CertificatesAggregator, VotesAggregator};
+use crate::error::{DagError, DagResult};
+use crate::messages::{Certificate, Header, Vote};
+use crate::primary::{PrimaryMessage, Round};
+use crate::synchronizer::Synchronizer;
+use async_recursion::async_recursion;
+use bytes::Bytes;
+use config::Committee;
+use crypto::Hash as _;
+use crypto::{Digest, PublicKey, SignatureService};
+use log::{debug, error, warn};
+use network::{CancelHandler, ReliableSender};
+use std::collections::{HashMap, HashSet};
+use std::sync::atomic::{AtomicU64, Ordering};
+use std::sync::Arc;
+use store::Store;
+use tokio::sync::mpsc::{Receiver, Sender};
+
+pub struct Core {
+    /// The public key of this primary.
+    name: PublicKey,
+    /// The committee information.
+    committee: Committee,
+    /// The persistent storage.
+    store: Store,
+    /// Handles synchronization with other nodes and our workers.
+    synchronizer: Synchronizer,
+    /// Service to sign headers.
+    signature_service: SignatureService,
+    /// The current consensus round (used for cleanup).
+    consensus_round: Arc<AtomicU64>,
+    /// The depth of the garbage collector.
+    gc_depth: Round,
+
+    /// Receiver for dag messages (headers, votes, certificates).
+    rx_primaries: Receiver<PrimaryMessage>,
+    /// Receives loopback messages from the `Waiter`.
+    rx_waiter: Receiver<Header>,
+    /// Receives our newly created headers from the `Proposer`.
+    rx_proposer: Receiver<Header>,
+    /// Output all certificates to the consensus layer.
+    tx_consensus: Sender<Certificate>,
+    /// Send valid a quorum of certificates' ids to the `Proposer` (along with their round).
+    tx_proposer: Sender<(Vec<Digest>, Round)>,
+
+    /// The last garbage collected round.
+    gc_round: Round,
+    /// The authors of the last voted headers.
+    last_voted: HashMap<Round, HashSet<PublicKey>>,
+    /// The set of headers we are currently processing.
+    processing: HashMap<Round, HashSet<Digest>>,
+    /// The last header we proposed (for which we are waiting votes).
+    current_header: Header,
+    /// Aggregates votes into a certificate.
+    votes_aggregator: VotesAggregator,
+    /// Aggregates certificates to use as parents for new headers.
+    certificates_aggregators: HashMap<Round, Box<CertificatesAggregator>>,
+    /// A network sender to send the batches to the other workers.
+    network: ReliableSender,
+    /// Keeps the cancel handlers of the messages we sent.
+    cancel_handlers: HashMap<Round, CancelHandler>,
+}
+
+impl Core {
+    pub fn spawn(
+        name: PublicKey,
+        committee: Committee,
+        store: Store,
+        synchronizer: Synchronizer,
+        signature_service: SignatureService,
+        consensus_round: Arc<AtomicU64>,
+        gc_depth: Round,
+        rx_primaries: Receiver<PrimaryMessage>,
+        rx_waiter: Receiver<Header>,
+        rx_proposer: Receiver<Header>,
+        tx_consensus: Sender<Certificate>,
+        tx_proposer: Sender<(Vec<Digest>, Round)>,
+    ) {
+        tokio::spawn(async move {
+            Self {
+                name,
+                committee,
+                store,
+                synchronizer,
+                signature_service,
+                consensus_round,
+                gc_depth,
+                rx_primaries,
+                rx_waiter,
+                rx_proposer,
+                tx_consensus,
+                tx_proposer,
+                gc_round: 0,
+                last_voted: HashMap::with_capacity(2 * gc_depth as usize),
+                processing: HashMap::with_capacity(2 * gc_depth as usize),
+                current_header: Header::default(),
+                votes_aggregator: VotesAggregator::new(),
+                certificates_aggregators: HashMap::with_capacity(2 * gc_depth as usize),
+                network: ReliableSender::new(),
+                cancel_handlers: HashMap::with_capacity(2 * gc_depth as usize),
+            }
+            .run()
+            .await;
+        });
+    }
+
+    async fn process_own_header(&mut self, header: &Header) -> DagResult<()> {
+        // Reset the votes aggregator.
+        self.current_header = header.clone();
+        self.votes_aggregator = VotesAggregator::new();
+
+        // Broadcast the new header in a reliable manner.
+        let addresses = self
+            .committee
+            .others_primaries(&self.name)
+            .iter()
+            .map(|(_, x)| x.primary_to_primary)
+            .collect();
+        let bytes = bincode::serialize(header).expect("Failed to serialize our own header");
+        let handlers = self.network.broadcast(addresses, Bytes::from(bytes)).await;
+        for handler in handlers {
+            self.cancel_handlers.insert(header.round, handler);
+        }
+
+        // Process the header.
+        self.process_header(&header).await
+    }
+
+    #[async_recursion]
+    async fn process_header(&mut self, header: &Header) -> DagResult<()> {
+        debug!("Processing {}", header);
+        // Indicate that we are processing this header.
+        self.processing
+            .entry(header.round)
+            .or_insert_with(HashSet::new)
+            .insert(header.id.clone());
+
+        // Ensure we have the parents. If at least one parent is missing, the synchronizer returns an empty
+        // vector; it will gather the missing parents (as well as all ancestors) from other nodes and then
+        // reschedule processing of this header.
+        let parents = self.synchronizer.get_parents(header).await?;
+        if parents.is_empty() {
+            debug!("Processing of {} suspended: missing payload", header.id);
+            return Ok(());
+        }
+
+        // Check the parent certificates. Ensure the parents form a quorum and are all from the previous round.
+        let mut stake = 0;
+        for x in parents {
+            ensure!(
+                x.round + 1 == header.round,
+                DagError::MalformedHeader(header.id.clone())
+            );
+            stake += self.committee.stake(&x.origin);
+        }
+        ensure!(
+            stake >= self.committee.quorum_threshold(),
+            DagError::HeaderRequiresQuorum(header.id.clone())
+        );
+
+        // Store the header.
+        let bytes = bincode::serialize(header).expect("Failed to serialize header");
+        self.store.write(header.id.to_vec(), bytes).await;
+
+        // Check if we can vote for this header.
+        if self
+            .last_voted
+            .entry(header.round)
+            .or_insert_with(HashSet::new)
+            .insert(header.author)
+        {
+            // Make a vote and send it to the header's creator.
+            let round = header.round;
+            let vote = Vote::new(header, &self.name, &mut self.signature_service).await;
+            if vote.origin == self.name {
+                self.process_vote(vote)
+                    .await
+                    .expect("Failed to process our own vote");
+            } else {
+                let address = self
+                    .committee
+                    .primary(&header.author)
+                    .expect("Author of valid header is not in the committee")
+                    .primary_to_primary;
+                let bytes = bincode::serialize(&PrimaryMessage::Vote(vote))
+                    .expect("Failed to serialize our own vote");
+                let handler = self.network.send(address, Bytes::from(bytes)).await;
+                self.cancel_handlers.insert(round, handler);
+            }
+        }
+        Ok(())
+    }
+
+    async fn handle_header(&mut self, header: &Header) -> DagResult<()> {
+        if self.gc_round > header.round {
+            return Ok(());
+        }
+
+        // TODO [issue #202]: Prevent bad nodes from sending junk headers with high round numbers.
+
+        // Verify the header's signature.
+        header.verify(&self.committee)?;
+
+        // Ensure we have the payload. If we don't, the synchronizer will ask our workers to get it, and then
+        // reschedule processing of this header once we have it.
+        if self.synchronizer.missing_payload(header).await? {
+            debug!("Processing of {} suspended: missing payload", header);
+            // TODO: schedule re-processing.
+            return Ok(());
+        }
+
+        // Process the header if all checks pass.
+        self.process_header(header).await
+    }
+
+    #[async_recursion]
+    async fn process_vote(&mut self, vote: Vote) -> DagResult<()> {
+        debug!("Processing {:?}", vote);
+
+        // Add it to the votes' aggregator and try to make a new certificate.
+        if let Some(certificate) =
+            self.votes_aggregator
+                .append(vote, &self.committee, &self.current_header)?
+        {
+            debug!("Assembled {:?}", certificate);
+
+            // Broadcast the certificate.
+            let addresses = self
+                .committee
+                .others_primaries(&self.name)
+                .iter()
+                .map(|(_, x)| x.primary_to_primary)
+                .collect();
+            let bytes = bincode::serialize(&PrimaryMessage::Certificate(certificate.clone()))
+                .expect("Failed to serialize our own certificate");
+            let handlers = self.network.broadcast(addresses, Bytes::from(bytes)).await;
+            for handler in handlers {
+                self.cancel_handlers.insert(certificate.round, handler);
+            }
+
+            // Process the new certificate.
+            self.process_certificate(certificate)
+                .await
+                .expect("Failed to process valid certificate");
+        }
+        Ok(())
+    }
+
+    async fn handle_vote(&mut self, vote: Vote) -> DagResult<()> {
+        if self.current_header.round > vote.round {
+            return Ok(());
+        }
+
+        // Ensure we receive a vote on the expected header.
+        ensure!(
+            vote.id == self.current_header.id
+                && vote.origin == self.name
+                && vote.round == self.current_header.round,
+            DagError::UnexpectedVote(vote.id)
+        );
+
+        // Verify the vote.
+        vote.verify(&self.committee)?;
+
+        // Process the vote if all checks pass.
+        self.process_vote(vote).await
+    }
+
+    #[async_recursion]
+    async fn process_certificate(&mut self, certificate: Certificate) -> DagResult<()> {
+        debug!("Processing {:?}", certificate);
+
+        // Process the header embedded in the certificate if we haven't already voted for it (if we already
+        // voted, it means we already processed it). Since this header got certified, we are sure that all
+        // the data it refers to (ie. its payload and its parents) are available. We can thus continue the
+        // processing of the certificate even if we don't have them in store right now.
+        if !self
+            .processing
+            .get(&certificate.header.round)
+            .map_or_else(|| false, |x| x.contains(&certificate.header.id))
+        {
+            // This function may still throw an error if the header is malformed.
+            self.handle_header(&certificate.header).await?;
+        }
+
+        // Store the certificate.
+        let bytes = bincode::serialize(&certificate).expect("Failed to serialize certificate");
+        self.store.write(certificate.digest().to_vec(), bytes).await;
+
+        // Check if we have enough certificates to enter a new dag round and propose a header.
+        if let Some(parents) = self
+            .certificates_aggregators
+            .entry(certificate.round)
+            .or_insert_with(|| Box::new(CertificatesAggregator::new()))
+            .append(certificate.clone(), &self.committee)?
+        {
+            // Send it to the `Proposer`.
+            self.tx_proposer
+                .send((parents, certificate.round))
+                .await
+                .expect("Failed to send certificate");
+        }
+
+        // Send it to the consensus layer.
+        let id = certificate.header.id.clone();
+        if let Err(e) = self.tx_consensus.send(certificate).await {
+            warn!(
+                "Failed to deliver certificate {} to the consensus: {}",
+                id, e
+            );
+        }
+        Ok(())
+    }
+
+    async fn handle_certificate(&mut self, certificate: Certificate) -> DagResult<()> {
+        if self.gc_round > certificate.round {
+            return Ok(());
+        }
+
+        // Verify the certificate.
+        certificate.verify(&self.committee)?;
+
+        // Process the certificate if all checks pass.
+        self.process_certificate(certificate).await
+    }
+
+    // Main loop listening to incoming messages.
+    pub async fn run(&mut self) {
+        loop {
+            let result = tokio::select! {
+                // We receive here messages from other primaries.
+                Some(message) = self.rx_primaries.recv() => {
+                    match message {
+                        PrimaryMessage::Header(header) => self.process_header(&header).await,
+                        PrimaryMessage::Vote(vote) => self.handle_vote(vote).await,
+                        PrimaryMessage::Certificate(certificate) => self.handle_certificate(certificate).await,
+                        _ => panic!("Unexpected core message")
+                    }
+                },
+
+                // We receive here loopback messages from the `Synchronizer`. Those are message for which we interrupted
+                // execution (we were missing some of their dependencies) and we are now ready to resume processing.
+                Some(header) = self.rx_waiter.recv() => self.process_header(&header).await,
+
+                // We also receive here our new headers created by the `Proposer`.
+                Some(header) = self.rx_proposer.recv() => self.process_own_header(&header).await,
+            };
+            match result {
+                Ok(()) => (),
+                Err(DagError::StoreError(e)) => error!("{}", e),
+                Err(e) => warn!("{}", e),
+            }
+
+            // Cleanup internal state.
+            let round = self.consensus_round.load(Ordering::Relaxed);
+            if round > self.gc_depth {
+                let gc_round = round - self.gc_depth;
+                self.last_voted.retain(|k, _| k >= &gc_round);
+                self.processing.retain(|k, _| k >= &gc_round);
+                self.certificates_aggregators.retain(|k, _| k >= &gc_round);
+                self.cancel_handlers.retain(|k, _| k >= &gc_round);
+                self.gc_round = gc_round;
+            }
+        }
+    }
+}
