+++ rust/primary/src/lib.rs
@@ -7,16 +7,27 @@ extern crate ed25519_dalek;
 extern crate rstest;
 
 #[macro_use]
-pub mod error;
+mod error;
+mod aggregators;
+mod core;
+mod helper;
+mod messages;
+mod payload_receiver;
+mod primary;
+mod proposer;
+mod synchronizer;
+mod waiter;
+
 pub mod committee;
 pub mod manage_primary;
-pub mod messages;
-pub mod primary;
+pub mod old_messages;
+pub mod old_primary;
+pub mod old_synchronizer;
 pub mod primary_net;
 pub mod processor;
 mod sync_driver;
 pub mod sync_server;
-pub mod synchronizer;
 pub mod types;
 
-pub use crate::messages::{PrimaryWorkerMessage, WorkerPrimaryMessage};
+pub use crate::messages::Certificate;
+pub use crate::primary::{Primary, PrimaryWorkerMessage, Round, WorkerPrimaryMessage};
