+++ rust/primary/src/proposer.rs
@@ -1,0 +1,131 @@
+use crate::messages::Header;
+use crate::primary::Round;
+use config::WorkerId;
+use crypto::{Digest, PublicKey, SignatureService};
+use log::{debug, info};
+use tokio::sync::mpsc::{Receiver, Sender};
+use tokio::time::{sleep, Duration, Instant};
+
+/// The proposer creates new headers and send them to the core for broadcasting and further processing.
+pub struct Proposer {
+    /// The public key of this primary.
+    name: PublicKey,
+    /// Service to sign headers.
+    signature_service: SignatureService,
+    /// The size of the headers' payload.
+    header_size: usize,
+    /// The maximum delay to wait for batches' digests.
+    max_header_delay: u64,
+
+    /// Receives the parents to include in the next header (along with their round number).
+    rx_core: Receiver<(Vec<Digest>, Round)>,
+    /// Receives the batches' digests from our workers.
+    rx_workers: Receiver<(Digest, WorkerId)>,
+    /// Sends newly created headers to the `Core`.
+    tx_core: Sender<Header>,
+
+    /// The current round of the dag.
+    round: Round,
+    /// Holds the certificates' ids waiting to be included in the next header.
+    last_parents: Vec<Digest>,
+    /// Holds the batches' digests waiting to be included in the next header.
+    digests: Vec<(Digest, WorkerId)>,
+    /// Keeps track of the size (in bytes) of batches' digests that we received so far.
+    payload_size: usize,
+}
+
+impl Proposer {
+    pub fn spawn(
+        name: PublicKey,
+        signature_service: SignatureService,
+        header_size: usize,
+        max_header_delay: u64,
+        rx_core: Receiver<(Vec<Digest>, Round)>,
+        rx_workers: Receiver<(Digest, WorkerId)>,
+        tx_core: Sender<Header>,
+    ) {
+        tokio::spawn(async move {
+            Self {
+                name,
+                signature_service,
+                header_size,
+                max_header_delay,
+                rx_core,
+                rx_workers,
+                tx_core,
+                round: 0,
+                last_parents: Vec::new(),
+                digests: Vec::with_capacity(2 * header_size),
+                payload_size: 0,
+            }
+            .run()
+            .await;
+        });
+    }
+
+    async fn make_header(&mut self) {
+        // Make a new header.
+        let header = Header::new(
+            self.name,
+            self.round,
+            self.digests.drain(..).collect(),
+            self.last_parents.drain(..).collect(),
+            &mut self.signature_service,
+        )
+        .await;
+        info!("Created {}", header);
+
+        // Send the new header to the `Core` that will broadcast and process it.
+        self.tx_core
+            .send(header)
+            .await
+            .expect("Failed to send header");
+    }
+
+    // Main loop listening to incoming messages.
+    pub async fn run(&mut self) {
+        let timer = sleep(Duration::from_millis(self.max_header_delay));
+        tokio::pin!(timer);
+
+        loop {
+            // Check if we can propose a new header. We propose a new header when one of the following
+            // conditions is met:
+            // 1. We have a quorum of certificates from the previous round and enough batches' digests;
+            // 2. We have a quorum of certificates from the previous round and the specified maximum
+            // inter-header delay has passed.
+            let enough_parents = !self.last_parents.is_empty();
+            let enough_digests = self.payload_size >= self.header_size;
+            let timer_expired = timer.is_elapsed();
+            if (enough_parents && enough_digests) || (enough_parents && timer_expired) {
+                // Make a new header.
+                self.make_header().await;
+
+                // Reschedule the timer.
+                let deadline = Instant::now() + Duration::from_millis(self.max_header_delay);
+                timer.as_mut().reset(deadline);
+            }
+
+            tokio::select! {
+                Some((parents, round)) = self.rx_core.recv() => {
+                    if round < self.round {
+                        continue;
+                    }
+
+                    // Advance to the next round.
+                    self.round = round + 1;
+                    debug!("Dag moved to round {}", self.round);
+
+                    // Signal that we have enough parent certificates to propose a new header.
+                    self.last_parents = parents;
+                }
+                Some((digest, worker_id)) = self.rx_workers.recv() => {
+                    self.payload_size += digest.size();
+                    self.digests.push((digest, worker_id));
+                }
+                () = &mut timer => {
+                    // Nothing to do.
+                }
+            }
+        }
+    }
+}
